Add pgaudit.log_statement setting
This setting, when turned off (not default), turns off all logging of
the full statement text and parameters.  This is analogous to
log_statement_once, except it never logs the statement.

Depending on requirements, the full statement text might not be
required in the audit log.  The combination of command tag and object
can be enough.  Omitting the full statement text makes the logs less
verbose and can also prevent some accidental information leaks.

RM18435